fix: make flat-link census audit sources complete
