+++ scripts/frontier_finite_flat_link_even_car_support_census_2026_07_23.py
@@ -1,52 +1,54 @@
 #!/usr/bin/env python3
-"""Self-contained finite certificate for a flat-link even-CAR support census.
+"""Finite certificate for a flat-link even-CAR support census.
 
 Paired note:
     docs/FINITE_FLAT_LINK_EVEN_CAR_SUPPORT_CENSUS_BOUNDED_THEOREM_NOTE_2026-07-23.md
 
-The runner imports no campaign module or receipt.  It reconstructs the finite
-graph/link code, the six-mode free-plus-contact factor family, the supplied
-topological chart, and every reported control from definitions in this file.
-It is a bounded support/census certificate, not a representation theorem,
-autonomous update law, or physical-site compiler.
+The complete computation is partitioned across this primary and two ordinary
+helper modules so every source remains byte-complete in the restricted audit
+packet.  It is a bounded support/census certificate, not a representation
+theorem, autonomous update law, or physical-site compiler.
 """
 
 from __future__ import annotations
 
-from collections import Counter, defaultdict
-from dataclasses import dataclass
 from hashlib import sha256
-from itertools import combinations, permutations, product
+from itertools import product
 import json
-import math
 from pathlib import Path
 
 import numpy as np
 from scipy.linalg import schur
 
+from frontier_finite_flat_link_even_car_support_census_graph_2026_07_23 import (
+    DIRECTIONS,
+    FRAMES,
+    REVERSE_MODE,
+    graph_link_code,
+)
+from frontier_finite_flat_link_even_car_support_census_factors_2026_07_23 import (
+    BETA,
+    CONTACT_COUPLING,
+    TOL,
+    common_coin,
+    covariance_controls,
+    factor_presentation,
+    local_factor_controls,
+    onsite_even_car_controls,
+    polynomial_controls,
+    three_mode_gate_controls,
+)
+
 
 ROOT = Path(__file__).resolve().parents[1]
 AUDIT_TIMEOUT_SEC = 1800
-TOL = 2e-11
-BETA = -0.3
-CONTACT_COUPLING = 0.37
-K = 129
+AUDIT_INPUT_PATHS = (
+    "scripts/frontier_finite_flat_link_even_car_support_census_graph_2026_07_23.py",
+    "scripts/frontier_finite_flat_link_even_car_support_census_factors_2026_07_23.py",
+)
 LENGTHS = (3, 6, 7)
 PASS = 0
 FAIL = 0
-
-DIRECTIONS = np.asarray(
-    ((1, 0, 0), (-1, 0, 0), (0, 1, 0), (0, -1, 0), (0, 0, 1), (0, 0, -1)),
-    dtype=int,
-)
-REVERSE_MODE = (1, 0, 3, 2, 5, 4)
-I6 = np.eye(6, dtype=complex)
-REVERSE = np.zeros((6, 6), dtype=complex)
-REVERSE[np.arange(6), REVERSE_MODE] = 1
-UNIFORM = np.ones(6, dtype=complex) / math.sqrt(6)
-P_SCALAR = np.outer(UNIFORM, UNIFORM.conj())
-P_EVEN = (I6 + REVERSE) / 2 - P_SCALAR
-P_VECTOR = (I6 - REVERSE) / 2
 
 
 def check(label: str, condition: bool, detail: object = "") -> None:
@@ -57,1698 +59,6 @@ def check(label: str, condition: bool, detail: object = "") -> None:
     else:
         FAIL += 1
         print(f"FAIL: {label} :: {detail}")
-
-
-def bit_indices(mask: int):
-    while mask:
-        bit = mask & -mask
-        yield bit.bit_length() - 1
-        mask ^= bit
-
-
-@dataclass(frozen=True)
-class Pauli:
-    """Pauli ``i^phase X^x Z^z`` represented by integer support masks."""
-
-    phase: int = 0
-    x: int = 0
-    z: int = 0
-
-    def __matmul__(self, other: "Pauli") -> "Pauli":
-        phase = (self.phase + other.phase + 2 * (self.z & other.x).bit_count()) % 4
-        return Pauli(phase, self.x ^ other.x, self.z ^ other.z)
-
-    def commutes(self, other: "Pauli") -> bool:
-        return ((self.x & other.z).bit_count() + (self.z & other.x).bit_count()) % 2 == 0
-
-    def symplectic(self, qubits: int) -> int:
-        return self.x | (self.z << qubits)
-
-
-def pauli_product(rows) -> Pauli:
-    result = Pauli()
-    for row in rows:
-        result = result @ row
-    return result
-
-
-def echelon(rows) -> dict[int, int]:
-    pivots: dict[int, int] = {}
-    for source in rows:
-        row = int(source)
-        while row:
-            pivot = row.bit_length() - 1
-            if pivot in pivots:
-                row ^= pivots[pivot]
-            else:
-                pivots[pivot] = row
-                break
-    return pivots
-
-
-def gf2_rank(rows) -> int:
-    return len(echelon(rows))
-
-
-def phase_rank(rows, qubits: int) -> tuple[int, int]:
-    pivots: dict[int, Pauli] = {}
-    inconsistent = 0
-    for original in rows:
-        row = original
-        vector = row.symplectic(qubits)
-        while vector:
-            pivot = vector.bit_length() - 1
-            if pivot in pivots:
-                row = row @ pivots[pivot]
-                vector = row.symplectic(qubits)
-            else:
-                pivots[pivot] = row
-                break
-        if not vector and row.phase % 4:
-            inconsistent += 1
-    return len(pivots), inconsistent
-
-
-def pauli_pivots(rows, qubits: int) -> dict[int, Pauli]:
-    """Return a phase-carrying echelon basis for a commuting Pauli family."""
-
-    pivots: dict[int, Pauli] = {}
-    for original in rows:
-        row = original
-        vector = row.symplectic(qubits)
-        while vector:
-            pivot = vector.bit_length() - 1
-            if pivot in pivots:
-                row = row @ pivots[pivot]
-                vector = row.symplectic(qubits)
-            else:
-                pivots[pivot] = row
-                break
-    return pivots
-
-
-def pauli_remainder(row: Pauli, pivots: dict[int, Pauli], qubits: int) -> Pauli:
-    """Reduce a Pauli including phase, so ``Pauli()`` means exact membership."""
-
-    vector = row.symplectic(qubits)
-    while vector:
-        pivot = vector.bit_length() - 1
-        if pivot not in pivots:
-            return row
-        row = row @ pivots[pivot]
-        vector = row.symplectic(qubits)
-    return row
-
-
-def quotient_complement(base, candidates) -> tuple[int, ...]:
-    pivots = echelon(base)
-    output = []
-    for source in candidates:
-        row = int(source)
-        while row:
-            pivot = row.bit_length() - 1
-            if pivot in pivots:
-                row ^= pivots[pivot]
-            else:
-                pivots[pivot] = row
-                output.append(int(source))
-                break
-    return tuple(output)
-
-
-def independent_pauli_basis(rows, qubits: int) -> tuple[Pauli, ...]:
-    pivots: dict[int, int] = {}
-    output = []
-    for pauli in rows:
-        reduced = pauli.symplectic(qubits)
-        while reduced:
-            pivot = reduced.bit_length() - 1
-            if pivot in pivots:
-                reduced ^= pivots[pivot]
-            else:
-                pivots[pivot] = reduced
-                output.append(pauli)
-                break
-    return tuple(output)
-
-
-def symplectic_product(left: int, right: int, qubits: int) -> int:
-    mask = (1 << qubits) - 1
-    return (
-        ((left & mask) & (right >> qubits)).bit_count()
-        + ((left >> qubits) & (right & mask)).bit_count()
-    ) & 1
-
-
-def symplectic_gram_rank(rows: tuple[int, ...], qubits: int) -> int:
-    gram = []
-    for left in rows:
-        row = 0
-        for index, right in enumerate(rows):
-            if symplectic_product(left, right, qubits):
-                row |= 1 << index
-        gram.append(row)
-    return gf2_rank(gram)
-
-
-def proper_cubic_frames() -> tuple[np.ndarray, ...]:
-    frames = []
-    for order in permutations(range(3)):
-        permutation = np.eye(3, dtype=int)[list(order)]
-        for signs in product((-1, 1), repeat=3):
-            frame = np.diag(signs) @ permutation
-            if round(np.linalg.det(frame)) == 1:
-                frames.append(frame)
-    unique = {tuple(int(value) for value in frame.ravel()): frame for frame in frames}
-    return tuple(unique[key] for key in sorted(unique))
-
-
-FRAMES = proper_cubic_frames()
-FRAME_INDEX = {tuple(int(value) for value in frame.ravel()): index for index, frame in enumerate(FRAMES)}
-
-
-def signed_axis(frame: np.ndarray, axis: int) -> tuple[int, int]:
-    image = frame @ np.eye(3, dtype=int)[:, axis]
-    target = int(np.flatnonzero(image)[0])
-    return target, int(image[target])
-
-
-def direction_map(frame: np.ndarray) -> tuple[int, ...]:
-    output = []
-    for direction in DIRECTIONS:
-        transformed = frame @ direction
-        output.append(int(np.flatnonzero(np.all(DIRECTIONS == transformed, axis=1))[0]))
-    return tuple(output)
-
-
-@dataclass(frozen=True)
-class BaseEdge:
-    u: int
-    v: int
-    kind: str
-    owner: tuple[int, int, int]
-
-
-class PyramidCellulation:
-    """Dual graph of the proper-cubic square-pyramid coarse-cell subdivision."""
-
-    def __init__(self, length: int):
-        if length < 3:
-            raise ValueError("periodic L>=3 required")
-        self.length = length
-        self.cells = tuple(product(range(length), repeat=3))
-        self.vertices = []
-        self.vertex_index = {}
-        for cell in self.cells:
-            for mode in range(6):
-                self.vertex_index[(cell, mode)] = len(self.vertices)
-                self.vertices.append((cell, mode))
-        self.edges: list[tuple[int, int, str, tuple[int, int, int]]] = []
-        self.edge_lookup: dict[frozenset[int], int] = {}
-
-        def add_edge(u: int, v: int, kind: str, owner) -> None:
-            key = frozenset((u, v))
-            if key in self.edge_lookup:
-                raise ValueError(("duplicate edge", u, v))
-            self.edge_lookup[key] = len(self.edges)
-            self.edges.append((u, v, kind, owner))
-
-        for cell in self.cells:
-            for left, right in combinations(range(6), 2):
-                if REVERSE_MODE[left] != right:
-                    add_edge(
-                        self.vertex_index[(cell, left)],
-                        self.vertex_index[(cell, right)],
-                        "internal_triangle",
-                        cell,
-                    )
-            for axis in range(3):
-                target = list(cell)
-                target[axis] = (target[axis] + 1) % length
-                add_edge(
-                    self.vertex_index[(cell, 2 * axis)],
-                    self.vertex_index[(tuple(target), 2 * axis + 1)],
-                    "outer_square",
-                    cell,
-                )
-        self.incident = [[] for _ in self.vertices]
-        for edge, (u, v, _kind, _owner) in enumerate(self.edges):
-            self.incident[u].append(edge)
-            self.incident[v].append(edge)
-        for row in self.incident:
-            row.sort()
-
-    def edge_between(self, u: int, v: int) -> int:
-        return self.edge_lookup[frozenset((u, v))]
-
-    def cycle_mask(self, vertices: list[int]) -> int:
-        mask = 0
-        for index, source in enumerate(vertices):
-            mask ^= 1 << self.edge_between(source, vertices[(index + 1) % len(vertices)])
-        return mask
-
-
-def primal_edge_cycles(graph: PyramidCellulation):
-    rows = []
-    for cell in graph.cells:
-        for bits in product((0, 1), repeat=3):
-            modes = [2 * axis + (0 if bits[axis] else 1) for axis in range(3)]
-            vertices = [graph.vertex_index[(cell, mode)] for mode in modes]
-            rows.append((graph.cycle_mask(vertices), vertices, "center_corner_edge"))
-    for corner in graph.cells:
-        for axis in range(3):
-            first, second = [value for value in range(3) if value != axis]
-
-            def local_cell(first_shift: int, second_shift: int):
-                cell = list(corner)
-                cell[first] = (cell[first] - first_shift) % graph.length
-                cell[second] = (cell[second] - second_shift) % graph.length
-                return tuple(cell)
-
-            def first_direction(shift: int) -> int:
-                return 2 * first + (0 if shift else 1)
-
-            def second_direction(shift: int) -> int:
-                return 2 * second + (0 if shift else 1)
-
-            labels = (
-                (local_cell(0, 0), first_direction(0)),
-                (local_cell(0, 0), second_direction(0)),
-                (local_cell(0, 1), second_direction(1)),
-                (local_cell(0, 1), first_direction(0)),
-                (local_cell(1, 1), first_direction(1)),
-                (local_cell(1, 1), second_direction(1)),
-                (local_cell(1, 0), second_direction(0)),
-                (local_cell(1, 0), first_direction(1)),
-            )
-            vertices = [graph.vertex_index[label] for label in labels]
-            rows.append((graph.cycle_mask(vertices), vertices, "coarse_grid_edge"))
-    return rows
-
-
-def wilson_cycles(graph: PyramidCellulation) -> list[list[int]]:
-    rows = []
-    for axis in range(3):
-        transverse = (axis + 1) % 3
-        vertices = []
-        for step in range(graph.length):
-            cell = [0, 0, 0]
-            cell[axis] = step
-            next_cell = list(cell)
-            next_cell[axis] = (next_cell[axis] + 1) % graph.length
-            vertices.extend(
-                (
-                    graph.vertex_index[(tuple(cell), 2 * axis)],
-                    graph.vertex_index[(tuple(next_cell), 2 * axis + 1)],
-                    graph.vertex_index[(tuple(next_cell), 2 * transverse)],
-                )
-            )
-        rows.append(vertices)
-    return rows
-
-
-@dataclass(frozen=True)
-class Edge:
-    u: int
-    v: int | None
-    kind: str
-    owner: tuple[int, int, int]
-    label: int = -1
-
-
-class PunctureGraph:
-    """One rough terminal and six sink spokes added to each coarse cell."""
-
-    def __init__(self, length: int):
-        self.length = length
-        self.base = PyramidCellulation(length)
-        self.cells = self.base.cells
-        self.cell_index = {cell: index for index, cell in enumerate(self.cells)}
-        self.matter_count = len(self.base.vertices)
-        self.vertices = list(self.base.vertices) + [(cell, "sink") for cell in self.cells]
-        self.sink_index = {cell: self.matter_count + self.cell_index[cell] for cell in self.cells}
-        self.edges: list[Edge] = []
-        self.edge_lookup = {}
-        self.spoke_lookup = {}
-        self.terminal_lookup = {}
-
-        def add_internal(u: int, v: int, kind: str, owner, label: int = -1) -> int:
-            key = frozenset((u, v))
-            if key in self.edge_lookup:
-                raise ValueError(("duplicate", u, v, kind))
-            index = len(self.edges)
-            self.edges.append(Edge(u, v, kind, owner, label))
-            self.edge_lookup[key] = index
-            return index
-
-        for u, v, kind, owner in self.base.edges:
-            add_internal(u, v, f"matter_{kind}", owner)
-        for cell in self.cells:
-            sink = self.sink_index[cell]
-            for mode in range(6):
-                index = add_internal(
-                    sink,
-                    self.base.vertex_index[(cell, mode)],
-                    "puncture_spoke",
-                    cell,
-                    mode,
-                )
-                self.spoke_lookup[(cell, mode)] = index
-        for cell in self.cells:
-            index = len(self.edges)
-            self.edges.append(Edge(self.sink_index[cell], None, "rough_terminal", cell, 0))
-            self.terminal_lookup[(cell, 0)] = index
-
-        self.incident = [[] for _ in self.vertices]
-        for edge, row in enumerate(self.edges):
-            self.incident[row.u].append(edge)
-            if row.v is not None:
-                self.incident[row.v].append(edge)
-        for row in self.incident:
-            row.sort()
-
-    @property
-    def qubits(self) -> int:
-        return len(self.edges)
-
-    def edge_between(self, u: int, v: int) -> int:
-        return self.edge_lookup[frozenset((u, v))]
-
-    def B(self, vertex: int) -> Pauli:
-        z = 0
-        for edge in self.incident[vertex]:
-            z ^= 1 << edge
-        return Pauli(z=z)
-
-    def A(self, source: int, target: int) -> Pauli:
-        edge = self.edge_between(source, target)
-        z = 0
-        for vertex in (source, target):
-            for incident in self.incident[vertex]:
-                if incident == edge:
-                    break
-                z ^= 1 << incident
-        return Pauli(0 if source < target else 2, 1 << edge, z)
-
-    def cycle_mask(self, vertices: list[int]) -> int:
-        mask = 0
-        for index, source in enumerate(vertices):
-            mask ^= 1 << self.edge_between(source, vertices[(index + 1) % len(vertices)])
-        return mask
-
-    def loop_pauli(self, vertices: list[int]) -> Pauli:
-        result = Pauli(phase=len(vertices) % 4)
-        for index, source in enumerate(vertices):
-            result = result @ self.A(source, vertices[(index + 1) % len(vertices)])
-        return result
-
-    def local_cycles(self):
-        rows = list(primal_edge_cycles(self.base))
-        for u, v, kind, owner in self.base.edges:
-            if kind == "internal_triangle":
-                vertices = [self.sink_index[owner], u, v]
-                rows.append((self.cycle_mask(vertices), vertices, "puncture_triangle"))
-        return rows
-
-    def wilson_cycles(self):
-        return wilson_cycles(self.base)
-
-    def cell_constraint(self, cell) -> Pauli:
-        result = self.B(self.sink_index[cell])
-        for mode in range(6):
-            result = result @ self.B(self.base.vertex_index[(cell, mode)])
-        return result
-
-    def mapped_matter_A(self, base_edge: int) -> Pauli:
-        u, v, kind, _owner = self.base.edges[base_edge]
-        result = self.A(u, v)
-        if kind == "outer_square":
-            left_cell = self.base.vertices[u][0]
-            right_cell = self.base.vertices[v][0]
-            result = result @ Pauli(
-                x=(1 << self.terminal_lookup[(left_cell, 0)])
-                ^ (1 << self.terminal_lookup[(right_cell, 0)])
-            )
-        return result
-
-
-def local_stabilizers(graph: PunctureGraph) -> tuple[Pauli, ...]:
-    return tuple(
-        [graph.loop_pauli(vertices) for _mask, vertices, _kind in graph.local_cycles()]
-        + [graph.cell_constraint(cell) for cell in graph.cells]
-    )
-
-
-def wilson_initializers(graph: PunctureGraph) -> tuple[Pauli, ...]:
-    return tuple(graph.loop_pauli(vertices) for vertices in graph.wilson_cycles())
-
-
-def gauge_Z(graph: PunctureGraph, cell) -> Pauli:
-    terminal = graph.terminal_lookup[(cell, 0)]
-    return graph.B(graph.sink_index[cell]) @ Pauli(z=1 << terminal)
-
-
-def gauge_generators(graph: PunctureGraph):
-    return tuple(gauge_Z(graph, cell) for cell in graph.cells)
-
-
-def hermitian_normalize(pauli: Pauli) -> Pauli:
-    return Pauli((pauli.x & pauli.z).bit_count() & 1, pauli.x, pauli.z)
-
-
-def onsite_hopping(graph: PunctureGraph, cell, left: int, right: int) -> Pauli:
-    if left == right:
-        raise ValueError("onsite hopping endpoints must differ")
-    if left > right:
-        canonical = onsite_hopping(graph, cell, right, left)
-        return Pauli((canonical.phase + 2) % 4, canonical.x, canonical.z)
-    source = graph.base.vertex_index[(cell, left)]
-    target = graph.base.vertex_index[(cell, right)]
-    direct_key = frozenset((source, target))
-    if direct_key in graph.base.edge_lookup:
-        return graph.mapped_matter_A(graph.base.edge_lookup[direct_key])
-    helper = next(
-        graph.base.vertex_index[(cell, mode)]
-        for mode in range(6)
-        if mode not in (left, right)
-        and frozenset((source, graph.base.vertex_index[(cell, mode)])) in graph.base.edge_lookup
-        and frozenset((graph.base.vertex_index[(cell, mode)], target)) in graph.base.edge_lookup
-    )
-    first = graph.mapped_matter_A(graph.base.edge_lookup[frozenset((source, helper))])
-    second = graph.mapped_matter_A(graph.base.edge_lookup[frozenset((helper, target))])
-    # The two-edge product already carries the intermediate Jordan-Wigner
-    # parity.  Multiplying B(helper) cancels that parity and breaks the even-CAR
-    # incidence algebra for the three absent reverse pairs.  The minus sign is
-    # the orientation calibration fixed by H_ij H_jk H_ik = -i on code.
-    result = hermitian_normalize(first @ second)
-    return Pauli((result.phase + 2) % 4, result.x, result.z)
-
-
-def add_cell(cell, axis: int, amount: int, length: int):
-    row = list(cell)
-    row[axis] = (row[axis] + amount) % length
-    return tuple(row)
-
-
-def link_code(length: int):
-    cells = tuple(product(range(length), repeat=3))
-    links = tuple((cell, axis) for cell in cells for axis in range(3))
-    index = {link: q for q, link in enumerate(links)}
-    stars = []
-    for cell in cells:
-        mask = 0
-        for axis in range(3):
-            mask ^= 1 << index[(cell, axis)]
-            mask ^= 1 << index[(add_cell(cell, axis, -1, length), axis)]
-        stars.append(Pauli(x=mask))
-    plaquettes = []
-    for cell in cells:
-        for first, second in combinations(range(3), 2):
-            mask = (
-                (1 << index[(cell, first)])
-                ^ (1 << index[(add_cell(cell, first, 1, length), second)])
-                ^ (1 << index[(add_cell(cell, second, 1, length), first)])
-                ^ (1 << index[(cell, second)])
-            )
-            plaquettes.append(Pauli(z=mask))
-    logical_z = []
-    logical_x = []
-    for axis in range(3):
-        logical_z.append(
-            Pauli(z=sum(1 << index[(tuple(step if value == axis else 0 for value in range(3)), axis)] for step in range(length)))
-        )
-        logical_x.append(
-            Pauli(x=sum(1 << index[(cell, axis)] for cell in cells if cell[axis] == 0))
-        )
-    return {
-        "cells": cells,
-        "links": links,
-        "index": index,
-        "qubits": len(links),
-        "stars": tuple(stars),
-        "plaquettes": tuple(plaquettes),
-        "logical_z": tuple(logical_z),
-        "logical_x": tuple(logical_x),
-    }
-
-
-def directed_link(left, right, length: int):
-    for axis in range(3):
-        if add_cell(left, axis, 1, length) == right:
-            return left, axis
-        if add_cell(right, axis, 1, length) == left:
-            return right, axis
-    raise AssertionError((left, right, length))
-
-
-def lift_link_z(graph_qubits: int, link_index: int) -> Pauli:
-    return Pauli(z=1 << (graph_qubits + link_index))
-
-
-def support(row: Pauli) -> int:
-    return row.x | row.z
-
-
-def periodic_l1(left, right, modulus: int) -> int:
-    return sum(min((a - b) % modulus, (b - a) % modulus) for a, b in zip(left, right))
-
-
-def support_diameter(mask: int, positions, modulus: int) -> int:
-    points = tuple(positions[q] for q in bit_indices(mask))
-    if len(points) < 2:
-        return 0
-    return max(periodic_l1(left, right, modulus) for left, right in combinations(points, 2))
-
-
-def graph_position(graph: PunctureGraph, qubit: int):
-    row = graph.edges[qubit]
-    center = 2 * K * np.asarray(row.owner, dtype=int)
-    if row.kind == "rough_terminal":
-        offset = np.zeros(3, dtype=int)
-    elif row.kind == "puncture_spoke":
-        offset = 8 * DIRECTIONS[row.label]
-    elif row.kind == "matter_internal_triangle":
-        left = graph.base.vertices[row.u][1]
-        right = graph.base.vertices[row.v][1]
-        offset = 4 * (DIRECTIONS[left] + DIRECTIONS[right])
-    elif row.kind == "matter_outer_square":
-        offset = 32 * DIRECTIONS[graph.base.vertices[row.u][1]]
-    else:
-        raise ValueError(row.kind)
-    return tuple(int(value % (2 * K * graph.length)) for value in center + offset)
-
-
-def link_midpoint(cell, axis: int, length: int):
-    modulus = 2 * K * length
-    row = [(2 * K * cell[value]) % modulus for value in range(3)]
-    row[axis] = (row[axis] + K) % modulus
-    return tuple(row)
-
-
-def sector_controls(code, length: int) -> dict:
-    """Execute the flat-link gradient/topological formula in all eight sectors."""
-
-    cells = code["cells"]
-    root = (0, 0, 0)
-    nonroot = tuple(cell for cell in cells if cell != root)
-    gauge_index = {cell: index for index, cell in enumerate(nonroot)}
-    topological_offset = len(nonroot)
-
-    def gauge_value(cell):
-        return 0 if cell == root else 1 << gauge_index[cell]
-
-    expressions = []
-    for cell, axis in code["links"]:
-        target = add_cell(cell, axis, 1, length)
-        value = gauge_value(cell) ^ gauge_value(target)
-        if cell[axis] == length - 1:
-            value ^= 1 << (topological_offset + axis)
-        expressions.append(value)
-
-    flatness_failures = 0
-    for plaquette in code["plaquettes"]:
-        value = 0
-        for q in bit_indices(plaquette.z):
-            value ^= expressions[q]
-        flatness_failures += value != 0
-
-    sector_failures = 0
-    sector_rows = []
-    for sector in range(8):
-        assignment = 0
-        for axis in range(3):
-            if (sector >> axis) & 1:
-                assignment |= 1 << (topological_offset + axis)
-        link_values = tuple((expression & assignment).bit_count() & 1 for expression in expressions)
-        plaquette_failures = sum(
-            (sum(link_values[q] for q in bit_indices(row.z)) & 1) != 0
-            for row in code["plaquettes"]
-        )
-        loop_failures = 0
-        measured = []
-        for axis in range(3):
-            cursor = [0, 0, 0]
-            value = 0
-            for _ in range(length):
-                value ^= link_values[code["index"][(tuple(cursor), axis)]]
-                cursor[axis] = (cursor[axis] + 1) % length
-            measured.append(value)
-            loop_failures += value != ((sector >> axis) & 1)
-        sector_failures += plaquette_failures + loop_failures
-        sector_rows.append(
-            {
-                "sector": tuple(1 if (sector >> axis) & 1 else 0 for axis in range(3)),
-                "measured_loop_bits": tuple(measured),
-                "plaquette_failures": plaquette_failures,
-                "loop_failures": loop_failures,
-            }
-        )
-
-    return {
-        "expression_rank": gf2_rank(expressions),
-        "expected_expression_rank": length**3 + 2,
-        "formula_flatness_failures": flatness_failures,
-        "all_eight_sector_failures": sector_failures,
-        "sector_rows": sector_rows,
-        "topological_input_qubits": 3,
-        "topological_input_preparation_or_genesis_constructed": False,
-        "pass": gf2_rank(expressions) == length**3 + 2
-        and flatness_failures == sector_failures == 0,
-    }
-
-
-def graph_link_code(length: int):
-    graph = PunctureGraph(length)
-    link = link_code(length)
-    cells = length**3
-    graph_qubits = graph.qubits
-    total = graph_qubits + link["qubits"]
-    local = local_stabilizers(graph)
-    wilsons = wilson_initializers(graph)
-    gauge_z = gauge_generators(graph)
-    cell_index = {cell: index for index, cell in enumerate(graph.cells)}
-
-    correlations = []
-    for link_index, (cell, axis) in enumerate(link["links"]):
-        target = add_cell(cell, axis, 1, length)
-        row = (
-            lift_link_z(graph_qubits, link_index)
-            @ gauge_z[cell_index[cell]]
-            @ gauge_z[cell_index[target]]
-        )
-        if cell[axis] == length - 1:
-            row = row @ wilsons[axis]
-        correlations.append(row)
-    correlations = tuple(correlations)
-    combined = local + correlations
-
-    lifted_plaquettes = []
-    plaquette_equal_failures = 0
-    for plaquette in link["plaquettes"]:
-        lifted = Pauli(plaquette.phase, plaquette.x << graph_qubits, plaquette.z << graph_qubits)
-        lifted_plaquettes.append(lifted)
-        actual = pauli_product(correlations[q] for q in bit_indices(plaquette.z))
-        plaquette_equal_failures += actual != lifted
-    lifted_plaquettes = tuple(lifted_plaquettes)
-
-    B = tuple(graph.B(vertex) for vertex in range(graph.matter_count))
-    A = []
-    A_link = []
-    outer = 0
-    for edge, (source, target, kind, _owner) in enumerate(graph.base.edges):
-        row = graph.mapped_matter_A(edge)
-        linked = None
-        if kind == "outer_square":
-            left = graph.base.vertices[source][0]
-            right = graph.base.vertices[target][0]
-            cell, axis = directed_link(left, right, length)
-            linked = link["index"][(cell, axis)]
-            row = row @ lift_link_z(graph_qubits, linked)
-            outer += 1
-        A.append(row)
-        A_link.append(linked)
-    A = tuple(A)
-    matter = B + A
-
-    local_rank, local_inconsistent = phase_rank(local, total)
-    combined_rank, combined_inconsistent = phase_rank(combined, total)
-    combined_vectors = tuple(row.symplectic(total) for row in combined)
-    matter_vectors = tuple(row.symplectic(total) for row in matter)
-    representatives = quotient_complement(combined_vectors, matter_vectors)
-    quotient_dimension = len(representatives)
-    quotient_gram_rank = symplectic_gram_rank(representatives, total)
-    stabilizer_commutator_failures = sum(
-        not generator.commutes(stabilizer)
-        for generator in matter
-        for stabilizer in combined
-    )
-
-    dressing_type_failures = 0
-    for edge, linked in enumerate(A_link):
-        delta = graph.mapped_matter_A(edge) @ A[edge]
-        expected = Pauli() if linked is None else lift_link_z(graph_qubits, linked)
-        dressing_type_failures += delta != expected
-
-    positions = tuple(graph_position(graph, q) for q in range(graph_qubits))
-    positions += tuple(link_midpoint(cell, axis, length) for cell, axis in link["links"])
-    modulus = 2 * K * length
-    placement_collisions = len(positions) - len(set(positions))
-
-    gauss = []
-    for cell in graph.cells:
-        parity = pauli_product(
-            B[graph.base.vertex_index[(cell, mode)]] for mode in range(6)
-        )
-        star_x = Pauli()
-        for axis in range(3):
-            star_x = star_x @ Pauli(x=1 << (graph_qubits + link["index"][(cell, axis)]))
-            prior = add_cell(cell, axis, -1, length)
-            star_x = star_x @ Pauli(x=1 << (graph_qubits + link["index"][(prior, axis)]))
-        gauss.append(parity @ star_x)
-    gauss = tuple(gauss)
-    gauss_matter_failures = sum(not row.commutes(generator) for row in gauss for generator in matter)
-    gauss_plaquette_failures = sum(not row.commutes(plaquette) for row in gauss for plaquette in lifted_plaquettes)
-
-    local_basis = independent_pauli_basis(local, total)
-    deleted_rank, _deleted_inconsistent = phase_rank(local_basis[1:] + correlations, total)
-    flipped = (
-        Pauli((correlations[0].phase + 2) % 4, correlations[0].x, correlations[0].z),
-    ) + correlations[1:]
-    _flipped_rank, flipped_inconsistent = phase_rank(local + flipped + (correlations[0],), total)
-    sectors = sector_controls(link, length)
-
-    maximum_gauss_diameter = max(
-        support_diameter(support(row), positions, modulus) for row in gauss
-    )
-    maximum_plaquette_diameter = max(
-        support_diameter(support(row), positions, modulus) for row in lifted_plaquettes
-    )
-    maximum_correlation_weight = max(support(row).bit_count() for row in correlations)
-    maximum_correlation_diameter = max(
-        support_diameter(support(row), positions, modulus) for row in correlations
-    )
-    row = {
-        "length": length,
-        "split": f"L{length}-finite-census",
-        "coarse_cells": cells,
-        "graph_M2": graph_qubits,
-        "flat_link_M2": link["qubits"],
-        "active_M2": total,
-        "active_M2_per_cell": total // cells,
-        "graph_local_rank": local_rank,
-        "expected_graph_local_rank": 15 * cells - 2,
-        "combined_rank": combined_rank,
-        "expected_combined_rank": 18 * cells - 2,
-        "combined_code_exponent": total - combined_rank,
-        "expected_code_exponent": 7 * cells + 2,
-        "matter_quotient_dimension": quotient_dimension,
-        "expected_matter_quotient_dimension": 12 * cells - 1,
-        "matter_quotient_symplectic_rank": quotient_gram_rank,
-        "expected_matter_quotient_symplectic_rank": 12 * cells - 2,
-        "matter_center_dimension": quotient_dimension - quotient_gram_rank,
-        "outer_link_dressed_A": outer,
-        "plaquette_product_equality_failures": plaquette_equal_failures,
-        "dressing_type_failures": dressing_type_failures,
-        "matter_constraint_commutator_failures": stabilizer_commutator_failures,
-        "local_Gauss_rows": len(gauss),
-        "Gauss_matter_commutator_failures": gauss_matter_failures,
-        "Gauss_plaquette_commutator_failures": gauss_plaquette_failures,
-        "maximum_Gauss_weight": max(support(row).bit_count() for row in gauss),
-        "maximum_plaquette_weight": max(support(row).bit_count() for row in lifted_plaquettes),
-        "maximum_Gauss_fine_L1_diameter": maximum_gauss_diameter,
-        "maximum_plaquette_fine_L1_diameter": maximum_plaquette_diameter,
-        "maximum_compile_time_correlation_weight": maximum_correlation_weight,
-        "maximum_compile_time_correlation_fine_L1_diameter": maximum_correlation_diameter,
-        "placement_collisions": placement_collisions,
-        "delete_one_independent_check_rank": deleted_rank,
-        "expected_deleted_rank": combined_rank - 1,
-        "malformed_phase_inconsistencies": flipped_inconsistent,
-        "topological_sector_controls": sectors,
-    }
-    row["pass"] = bool(
-        local_inconsistent == combined_inconsistent == 0
-        and local_rank == 15 * cells - 2
-        and combined_rank == 18 * cells - 2
-        and total - combined_rank == 7 * cells + 2
-        and quotient_dimension == 12 * cells - 1
-        and quotient_gram_rank == 12 * cells - 2
-        and quotient_dimension - quotient_gram_rank == 1
-        and outer == 3 * cells
-        and plaquette_equal_failures == dressing_type_failures == stabilizer_commutator_failures == 0
-        and gauss_matter_failures == gauss_plaquette_failures == placement_collisions == 0
-        and maximum_gauss_diameter <= 4 * K
-        and maximum_plaquette_diameter <= 2 * K
-        and deleted_rank == combined_rank - 1
-        and flipped_inconsistent > 0
-        and sectors["pass"]
-    )
-    internal = {
-        "graph": graph,
-        "link": link,
-        "B": B,
-        "A": A,
-        "A_link": tuple(A_link),
-        "positions": positions,
-        "local_constraints": local,
-        "correlation_section": correlations,
-        "combined_constraints": combined,
-        "graph_qubits": graph_qubits,
-        "total_qubits": total,
-    }
-    return row, internal
-
-
-def permute_mask(mask: int, mapping) -> int:
-    output = 0
-    for source in bit_indices(mask):
-        output ^= 1 << mapping[source]
-    return output
-
-
-def gf2_basis(rows):
-    return echelon(rows)
-
-
-def in_span(value: int, pivots) -> bool:
-    while value:
-        pivot = value.bit_length() - 1
-        if pivot not in pivots:
-            return False
-        value ^= pivots[pivot]
-    return True
-
-
-def link_mapping(code, frame: np.ndarray, length: int):
-    mapping = []
-    for cell, axis in code["links"]:
-        target_axis, sign = signed_axis(frame, axis)
-        mapped = tuple(int(value) % length for value in frame @ np.asarray(cell, dtype=int))
-        if sign < 0:
-            mapped = add_cell(mapped, target_axis, -1, length)
-        mapping.append(code["index"][(mapped, target_axis)])
-    return tuple(mapping)
-
-
-def link_covariance(code, length: int) -> dict:
-    mappings = tuple(link_mapping(code, frame, length) for frame in FRAMES)
-    frame_bijection_failures = sum(len(set(mapping)) != code["qubits"] for mapping in mappings)
-    group_failures = 0
-    for left_index, left in enumerate(FRAMES):
-        for right_index, right in enumerate(FRAMES):
-            direct = mappings[FRAME_INDEX[tuple(int(value) for value in (left @ right).ravel())]]
-            group_failures += tuple(
-                mappings[left_index][mappings[right_index][q]] for q in range(code["qubits"])
-            ) != direct
-    star_masks = {row.x for row in code["stars"]}
-    plaquette_masks = {row.z for row in code["plaquettes"]}
-    plaquette_basis = gf2_basis(plaquette_masks)
-    star_basis = gf2_basis(star_masks)
-    constraint_failures = 0
-    homology_failures = 0
-    for frame, mapping in zip(FRAMES, mappings):
-        constraint_failures += sum(permute_mask(row.x, mapping) not in star_masks for row in code["stars"])
-        constraint_failures += sum(permute_mask(row.z, mapping) not in plaquette_masks for row in code["plaquettes"])
-        for axis in range(3):
-            target, _sign = signed_axis(frame, axis)
-            homology_failures += not in_span(
-                permute_mask(code["logical_z"][axis].z, mapping) ^ code["logical_z"][target].z,
-                plaquette_basis,
-            )
-            homology_failures += not in_span(
-                permute_mask(code["logical_x"][axis].x, mapping) ^ code["logical_x"][target].x,
-                star_basis,
-            )
-    return {
-        "proper_cubic_frames": len(FRAMES),
-        "frame_products": len(FRAMES) ** 2,
-        "frame_bijection_failures": frame_bijection_failures,
-        "all576_group_failures": group_failures,
-        "constraint_image_failures": constraint_failures,
-        "logical_homology_image_failures": homology_failures,
-        "pass": frame_bijection_failures == group_failures == constraint_failures == homology_failures == 0,
-    }
-
-
-def graph_frame_maps(graph: PunctureGraph, frame: np.ndarray):
-    modes = direction_map(frame)
-    vertex_map = []
-    for vertex in range(graph.matter_count):
-        cell, mode = graph.base.vertices[vertex]
-        target_cell = tuple(int(value % graph.length) for value in frame @ np.asarray(cell))
-        vertex_map.append(graph.base.vertex_index[(target_cell, modes[mode])])
-    for cell in graph.cells:
-        target_cell = tuple(int(value % graph.length) for value in frame @ np.asarray(cell))
-        vertex_map.append(graph.sink_index[target_cell])
-    edge_map = []
-    for row in graph.edges:
-        if row.v is None:
-            target_cell = tuple(int(value % graph.length) for value in frame @ np.asarray(row.owner))
-            edge_map.append(graph.terminal_lookup[(target_cell, row.label)])
-        else:
-            edge_map.append(graph.edge_between(vertex_map[row.u], vertex_map[row.v]))
-    return tuple(vertex_map), tuple(edge_map)
-
-
-def order_gauge(graph: PunctureGraph, vertex_map, edge_map):
-    toggles = [0] * graph.qubits
-    pairs = []
-    for source_vertex, target_vertex in enumerate(vertex_map):
-        pulled = [edge_map[edge] for edge in graph.incident[source_vertex]]
-        position = {edge: index for index, edge in enumerate(graph.incident[target_vertex])}
-        for index, left in enumerate(pulled):
-            for right in pulled[index + 1 :]:
-                if position[left] > position[right]:
-                    toggles[left] ^= 1 << right
-                    toggles[right] ^= 1 << left
-                    pairs.append((left, right))
-    return tuple(toggles), tuple(pairs)
-
-
-@dataclass(frozen=True)
-class FrameData:
-    vertex_map: tuple[int, ...]
-    edge_map: tuple[int, ...]
-    toggles: tuple[int, ...]
-    pairs: frozenset[frozenset[int]]
-    flips: int
-
-
-def transform_pauli(pauli: Pauli, data: FrameData) -> Pauli:
-    x = z = 0
-    for source in bit_indices(pauli.x):
-        x ^= 1 << data.edge_map[source]
-    for source in bit_indices(pauli.z):
-        z ^= 1 << data.edge_map[source]
-    phase = pauli.phase
-    xbits = tuple(bit_indices(x))
-    if len(xbits) > 1:
-        phase = (phase + 2 * sum(frozenset(pair) in data.pairs for pair in combinations(xbits, 2))) % 4
-    for edge in xbits:
-        z ^= data.toggles[edge]
-    phase = (phase + 2 * (x & data.flips).bit_count()) % 4
-    return Pauli(phase, x, z)
-
-
-def frame_data(graph: PunctureGraph, frame: np.ndarray) -> FrameData:
-    vertex_map, edge_map = graph_frame_maps(graph, frame)
-    toggles, pairs = order_gauge(graph, vertex_map, edge_map)
-    pair_set = frozenset(frozenset(pair) for pair in pairs)
-    provisional = FrameData(vertex_map, edge_map, toggles, pair_set, 0)
-    flips = 0
-    for source_edge, row in enumerate(graph.edges):
-        if row.v is None:
-            continue
-        transformed = transform_pauli(graph.A(row.u, row.v), provisional)
-        target = graph.A(vertex_map[row.u], vertex_map[row.v])
-        if (transformed.phase - target.phase) % 4 == 2:
-            flips ^= 1 << edge_map[source_edge]
-    return FrameData(vertex_map, edge_map, toggles, pair_set, flips)
-
-
-def transform_graph_link_pauli(
-    pauli: Pauli,
-    data: FrameData,
-    link_map: tuple[int, ...],
-    graph_qubits: int,
-) -> Pauli:
-    """Transport both graph and link factors, preserving the Pauli phase."""
-
-    graph_mask = (1 << graph_qubits) - 1
-    graph_part = transform_pauli(
-        Pauli(pauli.phase, pauli.x & graph_mask, pauli.z & graph_mask), data
-    )
-    link_x = permute_mask(pauli.x >> graph_qubits, link_map) << graph_qubits
-    link_z = permute_mask(pauli.z >> graph_qubits, link_map) << graph_qubits
-    return Pauli(graph_part.phase, graph_part.x | link_x, graph_part.z | link_z)
-
-
-def covariance_controls(rows, internals) -> dict:
-    mode_maps = tuple(direction_map(frame) for frame in FRAMES)
-    mode_group_failures = 0
-    for left_index, left in enumerate(FRAMES):
-        for right_index, right in enumerate(FRAMES):
-            direct = mode_maps[FRAME_INDEX[tuple(int(value) for value in (left @ right).ravel())]]
-            mode_group_failures += tuple(
-                mode_maps[left_index][mode_maps[right_index][mode]] for mode in range(6)
-            ) != direct
-
-    size_rows = []
-    total_failures = mode_group_failures
-    for row, code in zip(rows, internals):
-        length = row["length"]
-        graph = code["graph"]
-        link = code["link"]
-        link_maps = tuple(link_mapping(link, frame, length) for frame in FRAMES)
-        B_failures = A_failures = link_dressing_failures = 0
-        combined_covariance_executed = length == 3
-        combined_local_failures = combined_correlation_failures = 0
-        combined_pivots = (
-            pauli_pivots(code["combined_constraints"], code["total_qubits"])
-            if combined_covariance_executed
-            else {}
-        )
-        for frame, link_map in zip(FRAMES, link_maps):
-            data = frame_data(graph, frame)
-            for source, generator in enumerate(code["B"]):
-                B_failures += transform_pauli(generator, data) != code["B"][data.vertex_map[source]]
-            for edge, (source, target, _kind, _owner) in enumerate(graph.base.edges):
-                transformed = transform_pauli(graph.mapped_matter_A(edge), data)
-                target_edge = graph.base.edge_lookup[frozenset((data.vertex_map[source], data.vertex_map[target]))]
-                target_raw = graph.mapped_matter_A(target_edge)
-                target_source, target_target, _target_kind, _target_owner = graph.base.edges[target_edge]
-                expected_phase = 2 if (data.vertex_map[source], data.vertex_map[target]) == (target_target, target_source) else 0
-                expected = Pauli((target_raw.phase + expected_phase) % 4, target_raw.x, target_raw.z)
-                A_failures += transformed != expected
-                linked = code["A_link"][edge]
-                mapped_link = None if linked is None else link_map[linked]
-                link_dressing_failures += mapped_link != code["A_link"][target_edge]
-            if combined_covariance_executed:
-                combined_local_failures += sum(
-                    pauli_remainder(
-                        transform_graph_link_pauli(
-                            generator, data, link_map, code["graph_qubits"]
-                        ),
-                        combined_pivots,
-                        code["total_qubits"],
-                    )
-                    != Pauli()
-                    for generator in code["local_constraints"]
-                )
-                combined_correlation_failures += sum(
-                    pauli_remainder(
-                        transform_graph_link_pauli(
-                            generator, data, link_map, code["graph_qubits"]
-                        ),
-                        combined_pivots,
-                        code["total_qubits"],
-                    )
-                    != Pauli()
-                    for generator in code["correlation_section"]
-                )
-        link_control = link_covariance(link, length)
-        failures = B_failures + A_failures + link_dressing_failures + int(not link_control["pass"])
-        total_failures += failures
-        size_rows.append(
-            {
-                "length": length,
-                "B_generator_map_failures": B_failures,
-                "oriented_A_map_failures": A_failures,
-                "link_dressing_map_failures": link_dressing_failures,
-                "combined_code_space_covariance_executed": combined_covariance_executed,
-                "combined_local_constraint_span_failures": combined_local_failures if combined_covariance_executed else None,
-                "fixed_chart_correlation_span_failures": combined_correlation_failures if combined_covariance_executed else None,
-                "fixed_combined_code_space_covariant": (
-                    combined_local_failures == combined_correlation_failures == 0
-                    if combined_covariance_executed
-                    else None
-                ),
-                "transported_chart_covariant_by_construction": True,
-                "link_covariance": link_control,
-                "pass": failures == 0,
-            }
-        )
-    return {
-        "proper_cubic_frames": len(FRAMES),
-        "frame_products": len(FRAMES) ** 2,
-        "signed_six_mode_all576_group_failures": mode_group_failures,
-        "fixed_chart_combined_code_space_covariant_at_L3": size_rows[0]["fixed_combined_code_space_covariant"],
-        "combined_code_space_covariance_sizes": (3,),
-        "fixed_chart_invariant_claimed": False,
-        "compile_time_chart_transport_supplied": True,
-        "size_rows": size_rows,
-        "pass": total_failures == 0,
-    }
-
-
-def onsite_even_car_controls(internals) -> dict:
-    """Execute the complete 6 B / 15 H onsite incidence algebra and transport."""
-
-    size_rows = []
-    total_failures = 0
-    for code in internals:
-        graph = code["graph"]
-        local = code["local_constraints"]
-        pivots = pauli_pivots(local, graph.qubits)
-        B_square_failures = B_pair_commutator_failures = 0
-        H_square_failures = B_H_incidence_failures = H_H_incidence_failures = 0
-        H_triangle_on_code_phase_failures = 0
-        derived_exact_frame_failures = derived_on_code_frame_failures = 0
-
-        for cell in graph.cells:
-            B = tuple(
-                graph.B(graph.base.vertex_index[(cell, mode)]) for mode in range(6)
-            )
-            H = {
-                pair: onsite_hopping(graph, cell, *pair)
-                for pair in combinations(range(6), 2)
-            }
-            B_square_failures += sum(row @ row != Pauli() for row in B)
-            B_pair_commutator_failures += sum(
-                not B[left].commutes(B[right]) for left, right in combinations(range(6), 2)
-            )
-            H_square_failures += sum(row @ row != Pauli() for row in H.values())
-            B_H_incidence_failures += sum(
-                B[mode].commutes(row) == (mode in pair)
-                for pair, row in H.items()
-                for mode in range(6)
-            )
-            H_H_incidence_failures += sum(
-                left_row.commutes(right_row) == (len(set(left_pair) & set(right_pair)) == 1)
-                for (left_pair, left_row), (right_pair, right_row) in combinations(H.items(), 2)
-            )
-            H_triangle_on_code_phase_failures += sum(
-                pauli_remainder(
-                    H[(left, middle)] @ H[(middle, right)] @ H[(left, right)],
-                    pivots,
-                    graph.qubits,
-                )
-                != Pauli(phase=3)
-                for left, middle, right in combinations(range(6), 3)
-            )
-
-        for frame in FRAMES:
-            data = frame_data(graph, frame)
-            modes = direction_map(frame)
-            # One cell represents the translation orbit; the definitions and
-            # frame map are exactly translation-covariant on the periodic torus.
-            for cell in (graph.cells[0],):
-                target_cell = tuple(
-                    int(value % graph.length) for value in frame @ np.asarray(cell)
-                )
-                for left, right in combinations(range(6), 2):
-                    transformed = transform_pauli(
-                        onsite_hopping(graph, cell, left, right), data
-                    )
-                    mapped_left, mapped_right = modes[left], modes[right]
-                    target = onsite_hopping(
-                        graph,
-                        target_cell,
-                        min(mapped_left, mapped_right),
-                        max(mapped_left, mapped_right),
-                    )
-                    if mapped_left > mapped_right:
-                        target = Pauli((target.phase + 2) % 4, target.x, target.z)
-                    difference = transformed @ target
-                    derived_exact_frame_failures += difference != Pauli()
-                    derived_on_code_frame_failures += (
-                        pauli_remainder(difference, pivots, graph.qubits) != Pauli()
-                    )
-
-        failures = (
-            B_square_failures
-            + B_pair_commutator_failures
-            + H_square_failures
-            + B_H_incidence_failures
-            + H_H_incidence_failures
-            + H_triangle_on_code_phase_failures
-            + derived_on_code_frame_failures
-        )
-        total_failures += failures
-        size_rows.append(
-            {
-                "length": graph.length,
-                "onsite_B_rows": 6 * graph.length**3,
-                "onsite_H_rows": 15 * graph.length**3,
-                "B_square_failures": B_square_failures,
-                "B_pair_commutator_failures": B_pair_commutator_failures,
-                "H_square_failures": H_square_failures,
-                "B_H_endpoint_incidence_failures": B_H_incidence_failures,
-                "H_H_endpoint_incidence_failures": H_H_incidence_failures,
-                "H_triangle_minus_i_on_code_failures": H_triangle_on_code_phase_failures,
-                "derived_all24_exact_Pauli_failures": derived_exact_frame_failures,
-                "derived_all24_on_code_failures": derived_on_code_frame_failures,
-                "derived_frame_translation_orbit_representatives": 1,
-                "pass": failures == 0,
-            }
-        )
-    return {
-        "proper_cubic_frames": len(FRAMES),
-        "complete_onsite_bilinears_per_cell": 15,
-        "size_rows": size_rows,
-        "pass": total_failures == 0,
-    }
-
-
-@dataclass(frozen=True)
-class ModeGate:
-    kind: str
-    sites: tuple[int, ...]
-    matrix: tuple[complex, ...]
-
-
-def common_coin(beta: float = BETA) -> tuple[np.ndarray, float]:
-    inertial_mass = float(3 * np.tan(-beta / 2))
-    rest_phase = inertial_mass / 3
-    coin = np.exp(1j * rest_phase) * (
-        P_SCALAR - P_EVEN + np.exp(1j * beta) * P_VECTOR
-    )
-    return coin, inertial_mass
-
-
-def one_particle_matrix(gate: ModeGate) -> np.ndarray:
-    size = 1 if gate.kind == "phase" else 2
-    return np.asarray(gate.matrix, dtype=complex).reshape(size, size)
-
-
-def compile_adjacent_qr(unitary: np.ndarray):
-    work = unitary.copy()
-    eliminations = []
-    for column in range(5):
-        for lower in range(5, column, -1):
-            upper = lower - 1
-            a = work[upper, column]
-            b = work[lower, column]
-            if abs(b) < 1e-13:
-                continue
-            radius = np.sqrt(abs(a) ** 2 + abs(b) ** 2)
-            elimination = np.asarray(
-                ((np.conj(a) / radius, np.conj(b) / radius), (-b / radius, a / radius)),
-                dtype=complex,
-            )
-            work[[upper, lower], :] = elimination @ work[[upper, lower], :]
-            eliminations.append((upper, lower, elimination))
-    schedule = []
-    for index, phase in enumerate(np.diag(work)):
-        if abs(phase - 1) >= 1e-13:
-            schedule.append(ModeGate("phase", (index,), (complex(phase),)))
-    for upper, lower, elimination in reversed(eliminations):
-        schedule.append(ModeGate("givens", (upper, lower), tuple(elimination.conj().T.reshape(-1))))
-    reconstructed = np.eye(6, dtype=complex)
-    for gate in schedule:
-        factor = np.eye(6, dtype=complex)
-        factor[np.ix_(gate.sites, gate.sites)] = one_particle_matrix(gate)
-        reconstructed = factor @ reconstructed
-    return tuple(schedule), {
-        "givens": sum(gate.kind == "givens" for gate in schedule),
-        "phases": sum(gate.kind == "phase" for gate in schedule),
-        "diagonalization_residual": float(np.linalg.norm(work - np.diag(np.diag(work)))),
-        "reconstruction_residual": float(np.linalg.norm(reconstructed - unitary)),
-        "reconstructed": reconstructed,
-    }
-
-
-def occupied_modes(basis: int, mode_count: int):
-    return tuple(mode for mode in range(mode_count) if (basis >> mode) & 1)
-
-
-def fock_lift(unitary: np.ndarray) -> np.ndarray:
-    mode_count = unitary.shape[0]
-    dimension = 1 << mode_count
-    occupied = tuple(occupied_modes(basis, mode_count) for basis in range(dimension))
-    output = np.zeros((dimension, dimension), dtype=complex)
-    for target, target_modes in enumerate(occupied):
-        for source, source_modes in enumerate(occupied):
-            if len(target_modes) != len(source_modes):
-                continue
-            output[target, source] = 1 if not target_modes else np.linalg.det(
-                unitary[np.ix_(target_modes, source_modes)]
-            )
-    return output
-
-
-def embedded_gate(gate: ModeGate) -> np.ndarray:
-    one_particle = np.eye(6, dtype=complex)
-    one_particle[np.ix_(gate.sites, gate.sites)] = one_particle_matrix(gate)
-    return fock_lift(one_particle)
-
-
-def number_sector_leakage(matrix: np.ndarray) -> float:
-    numbers = np.asarray([basis.bit_count() for basis in range(matrix.shape[0])])
-    forbidden = numbers[:, None] != numbers[None, :]
-    return float(np.linalg.norm(matrix[forbidden]))
-
-
-def local_factor_controls() -> tuple[dict, tuple[ModeGate, ...]]:
-    coin, analytic_mass = common_coin()
-    schedule, qr = compile_adjacent_qr(coin)
-    gamma_coin = fock_lift(coin)
-    compiled_coin = np.eye(64, dtype=complex)
-    for gate in schedule:
-        compiled_coin = embedded_gate(gate) @ compiled_coin
-
-    contact_diagonal = np.asarray(
-        [np.exp(1j * CONTACT_COUPLING * basis.bit_count() * (basis.bit_count() - 1) / 2) for basis in range(64)],
-        dtype=complex,
-    )
-    contact = np.diag(contact_diagonal)
-    compiled_contact_diagonal = np.ones(64, dtype=complex)
-    for left, right in combinations(range(6), 2):
-        for basis in range(64):
-            if ((basis >> left) & 1) and ((basis >> right) & 1):
-                compiled_contact_diagonal[basis] *= np.exp(1j * CONTACT_COUPLING)
-    compiled_contact = np.diag(compiled_contact_diagonal)
-
-    reverse_gates = tuple(
-        ModeGate("givens", pair, tuple(np.asarray(((0, 1), (1, 0)), dtype=complex).reshape(-1)))
-        for pair in ((0, 1), (2, 3), (4, 5))
-    )
-    compiled_reverse = np.eye(64, dtype=complex)
-    for gate in reverse_gates:
-        compiled_reverse = embedded_gate(gate) @ compiled_reverse
-    direct_reverse = fock_lift(REVERSE)
-
-    # This is a canonical 64-dimensional word reconstruction.  It is not an
-    # intertwiner into the graph/link M2 factors: no such E is constructed here.
-    direct = contact @ direct_reverse @ gamma_coin
-    compiled = compiled_contact @ compiled_reverse @ compiled_coin
-    explicit_inverse = np.eye(64, dtype=complex)
-    for factor in (
-        compiled_contact.conj().T,
-        compiled_reverse.conj().T,
-        compiled_coin.conj().T,
-    ):
-        explicit_inverse = factor @ explicit_inverse
-
-    deleted_coin = np.eye(64, dtype=complex)
-    for gate in schedule[1:]:
-        deleted_coin = embedded_gate(gate) @ deleted_coin
-    deleted_contact_diagonal = np.ones(64, dtype=complex)
-    deleted_pair = next(iter(combinations(range(6), 2)))
-    for left, right in tuple(combinations(range(6), 2))[1:]:
-        for basis in range(64):
-            if ((basis >> left) & 1) and ((basis >> right) & 1):
-                deleted_contact_diagonal[basis] *= np.exp(1j * CONTACT_COUPLING)
-    deleted_contact = np.diag(deleted_contact_diagonal)
-
-    scalar_phase = complex(np.vdot(UNIFORM, qr["reconstructed"] @ UNIFORM))
-    compiled_mass = float(np.angle(scalar_phase)) / (1 / 3)
-    identity = np.eye(64, dtype=complex)
-    rows = {
-        "QR_Givens": qr["givens"],
-        "QR_onsite_phases": qr["phases"],
-        "one_particle_reconstruction_residual": qr["reconstruction_residual"],
-        "exterior_coin_reconstruction_residual": float(np.linalg.norm(compiled_coin - gamma_coin)),
-        "fifteen_contact_reconstruction_residual": float(np.linalg.norm(compiled_contact - contact)),
-        "reverse_FSWAP_word_reconstruction_residual": float(np.linalg.norm(compiled_reverse - direct_reverse)),
-        "full_M64_ordered_word_reconstruction_residual": float(np.linalg.norm(compiled - direct)),
-        "full_M64_word_unitarity_residual": float(np.linalg.norm(compiled.conj().T @ compiled - identity)),
-        "full_M64_explicit_inverse_residual": float(np.linalg.norm(explicit_inverse @ compiled - identity)),
-        "number_sector_leakage_residual": number_sector_leakage(compiled),
-        "compiled_rest_mass": compiled_mass,
-        "analytic_mass_fixture": analytic_mass,
-        "mass_fixture_residual": abs(compiled_mass - analytic_mass),
-        "contact_active_two_particle_states": sum(basis.bit_count() == 2 for basis in range(64)),
-        "contact_deletion_residual": float(np.linalg.norm(deleted_contact - contact, ord=2)),
-        "expected_contact_deletion_residual": float(abs(np.exp(1j * CONTACT_COUPLING) - 1)),
-        "deleted_coin_factor_residual": float(np.linalg.norm(deleted_coin - gamma_coin, ord=2)),
-        "deleted_contact_pair": deleted_pair,
-    }
-    rows["pass"] = bool(
-        rows["QR_Givens"] == 10
-        and rows["QR_onsite_phases"] == 1
-        and rows["one_particle_reconstruction_residual"] < TOL
-        and rows["exterior_coin_reconstruction_residual"] < TOL
-        and rows["fifteen_contact_reconstruction_residual"] < TOL
-        and rows["reverse_FSWAP_word_reconstruction_residual"] < TOL
-        and rows["full_M64_ordered_word_reconstruction_residual"] < TOL
-        and rows["full_M64_word_unitarity_residual"] < TOL
-        and rows["full_M64_explicit_inverse_residual"] < TOL
-        and rows["number_sector_leakage_residual"] == 0
-        and rows["mass_fixture_residual"] < TOL
-        and rows["contact_active_two_particle_states"] == 15
-        and abs(rows["contact_deletion_residual"] - rows["expected_contact_deletion_residual"]) < TOL
-        and rows["deleted_coin_factor_residual"] > 1e-3
-    )
-    return rows, schedule
-
-
-def polynomial_controls(schedule: tuple[ModeGate, ...]) -> dict:
-    eye = np.eye(2, dtype=complex)
-    x = np.asarray(((0, 1), (1, 0)), dtype=complex)
-    y = np.asarray(((0, -1j), (1j, 0)), dtype=complex)
-    z = np.diag((1, -1)).astype(complex)
-    left_B = np.kron(z, eye)
-    right_B = np.kron(eye, z)
-    hopping = np.kron(y, x)
-    basis = []
-    for left, right, active in product((0, 1), repeat=3):
-        basis.append(
-            (
-                f"Bl^{left} Br^{right} A^{active}",
-                np.linalg.matrix_power(left_B, left)
-                @ np.linalg.matrix_power(right_B, right)
-                @ np.linalg.matrix_power(hopping, active),
-            )
-        )
-
-    def expand(matrix):
-        coefficients = tuple(np.trace(row.conj().T @ matrix) / 4 for _name, row in basis)
-        rebuilt = sum(
-            (value * row for value, (_name, row) in zip(coefficients, basis)),
-            np.zeros((4, 4), dtype=complex),
-        )
-        return coefficients, float(np.linalg.norm(rebuilt - matrix))
-
-    maximum_coin = 0.0
-    payload = []
-    for gate in schedule:
-        if gate.kind == "givens":
-            matrix = fock_lift(one_particle_matrix(gate))
-            coefficients, residual = expand(matrix)
-            maximum_coin = max(maximum_coin, residual)
-            payload.append(
-                {
-                    "sites": gate.sites,
-                    "coefficients": tuple(
-                        (basis[index][0], value.real.hex(), value.imag.hex())
-                        for index, value in enumerate(coefficients)
-                    ),
-                }
-            )
-
-    phase = np.exp(1j * CONTACT_COUPLING)
-    contact = np.diag((1, 1, 1, phase)).astype(complex)
-    contact_coefficients, contact_residual = expand(contact)
-    fswap = fock_lift(np.asarray(((0, 1), (1, 0)), dtype=complex))
-    fswap_coefficients, fswap_residual = expand(fswap)
-    expected_fswap = {
-        "Bl^0 Br^1 A^0": 0.5,
-        "Bl^0 Br^1 A^1": -0.5j,
-        "Bl^1 Br^0 A^0": 0.5,
-        "Bl^1 Br^0 A^1": 0.5j,
-    }
-    fswap_sign_failures = sum(
-        abs(value - expected_fswap.get(name, 0)) > 1e-14
-        for value, (name, _row) in zip(fswap_coefficients, basis)
-    )
-    contact_expected = {
-        "Bl^0 Br^0 A^0": (3 + phase) / 4,
-        "Bl^0 Br^1 A^0": (1 - phase) / 4,
-        "Bl^1 Br^0 A^0": (1 - phase) / 4,
-        "Bl^1 Br^1 A^0": (phase - 1) / 4,
-    }
-    contact_sign_failures = sum(
-        abs(value - contact_expected.get(name, 0)) > 1e-14
-        for value, (name, _row) in zip(contact_coefficients, basis)
-    )
-    rows = {
-        "maximum_coin_polynomial_reconstruction_residual": maximum_coin,
-        "contact_polynomial_reconstruction_residual": contact_residual,
-        "FSWAP_polynomial_reconstruction_residual": fswap_residual,
-        "contact_exact_coefficient_sign_failures": contact_sign_failures,
-        "FSWAP_exact_coefficient_sign_failures": fswap_sign_failures,
-        "coin_coefficient_sha256": sha256(json.dumps(payload, sort_keys=True).encode()).hexdigest(),
-    }
-    rows["pass"] = bool(
-        maximum_coin < TOL
-        and contact_residual < TOL
-        and fswap_residual < TOL
-        and contact_sign_failures == fswap_sign_failures == 0
-    )
-    return rows
-
-
-def three_mode_gate_controls(schedule: tuple[ModeGate, ...]) -> dict:
-    """Catch cancellation of the helper parity in a two-edge onsite path.
-
-    The endpoint modes occupy positions 0 and 2 and the path helper occupies
-    position 1.  In little-endian Fock order the correct endpoint bilinear is
-    ``Y_2 Z_1 X_0``.  The former extra ``B_helper`` changed it to
-    ``Y_2 I_1 X_0``; applying identical polynomial coefficients to that wrong
-    bilinear gives an operator-level comparison control.
-    """
-
-    eye = np.eye(2, dtype=complex)
-    x = np.asarray(((0, 1), (1, 0)), dtype=complex)
-    y = np.asarray(((0, -1j), (1j, 0)), dtype=complex)
-    z = np.diag((1, -1)).astype(complex)
-    left_B = np.kron(np.kron(z, eye), eye)
-    helper_B = np.kron(np.kron(eye, z), eye)
-    right_B = np.kron(np.kron(eye, eye), z)
-    correct_A = np.kron(np.kron(y, z), x)
-    cancelled_A = helper_B @ correct_A
-
-    def algebra_basis(active_A):
-        return tuple(
-            np.linalg.matrix_power(left_B, left)
-            @ np.linalg.matrix_power(right_B, right)
-            @ np.linalg.matrix_power(active_A, active)
-            for left, right, active in product((0, 1), repeat=3)
-        )
-
-    correct_basis = algebra_basis(correct_A)
-    cancelled_basis = algebra_basis(cancelled_A)
-
-    def residuals(matrix):
-        coefficients = tuple(np.trace(row.conj().T @ matrix) / 8 for row in correct_basis)
-        correct = sum(
-            (coefficient * row for coefficient, row in zip(coefficients, correct_basis)),
-            np.zeros((8, 8), dtype=complex),
-        )
-        cancelled = sum(
-            (coefficient * row for coefficient, row in zip(coefficients, cancelled_basis)),
-            np.zeros((8, 8), dtype=complex),
-        )
-        return float(np.linalg.norm(correct - matrix)), float(np.linalg.norm(cancelled - matrix, ord=2))
-
-    reverse_pairs = {(0, 1), (2, 3), (4, 5)}
-    tested = []
-    for gate in schedule:
-        if gate.kind != "givens" or tuple(gate.sites) not in reverse_pairs:
-            continue
-        one_particle = np.eye(3, dtype=complex)
-        one_particle[np.ix_((0, 2), (0, 2))] = one_particle_matrix(gate)
-        tested.append(("coin_Givens", gate.sites, *residuals(fock_lift(one_particle))))
-    swap = np.eye(3, dtype=complex)
-    swap[np.ix_((0, 2), (0, 2))] = np.asarray(((0, 1), (1, 0)), dtype=complex)
-    for pair in sorted(reverse_pairs):
-        tested.append(("reverse_FSWAP", pair, *residuals(fock_lift(swap))))
-
-    correct = [row[2] for row in tested]
-    cancelled = [row[3] for row in tested]
-    rows = {
-        "three_mode_factors_tested": len(tested),
-        "correct_two_edge_maximum_Frobenius_residual": max(correct),
-        "extra_helper_B_minimum_operator_residual": min(cancelled),
-        "factor_rows": tuple(
-            {
-                "kind": kind,
-                "coarse_modes": pair,
-                "correct_Frobenius_residual": good,
-                "extra_helper_B_operator_residual": bad,
-            }
-            for kind, pair, good, bad in tested
-        ),
-    }
-    rows["pass"] = bool(
-        rows["three_mode_factors_tested"] == 8
-        and rows["correct_two_edge_maximum_Frobenius_residual"] < TOL
-        and rows["extra_helper_B_minimum_operator_residual"] > 1e-2
-    )
-    return rows
-
-
-def factor_presentation(length: int, code, schedule: tuple[ModeGate, ...]):
-    graph = code["graph"]
-    B = code["B"]
-    A = code["A"]
-    factors = []
-
-    def bv(cell, mode):
-        return B[graph.base.vertex_index[(cell, mode)]]
-
-    def add(kind, stage, cell, modes, rows):
-        mask = 0
-        for row in rows:
-            mask |= support(row)
-        factors.append(
-            {
-                "kind": kind,
-                "stage": stage,
-                "cell": cell,
-                "modes": modes,
-                "support": mask,
-                "weight": mask.bit_count(),
-            }
-        )
-
-    for gate_index, gate in enumerate(schedule):
-        for cell in graph.cells:
-            if gate.kind == "phase":
-                mode = gate.sites[0]
-                add("coin_phase", f"coin_{gate_index}", cell, (mode,), (bv(cell, mode),))
-            else:
-                left, right = gate.sites
-                add(
-                    "coin_Givens",
-                    f"coin_{gate_index}",
-                    cell,
-                    (left, right),
-                    (bv(cell, left), bv(cell, right), onsite_hopping(graph, cell, left, right)),
-                )
-    for reverse_index, (left, right) in enumerate(((0, 1), (2, 3), (4, 5))):
-        for cell in graph.cells:
-            add(
-                "reverse_FSWAP",
-                f"reverse_{reverse_index}",
-                cell,
-                (left, right),
-                (bv(cell, left), bv(cell, right), onsite_hopping(graph, cell, left, right)),
-            )
-    for edge, (source, target, kind, _owner) in enumerate(graph.base.edges):
-        if kind == "outer_square":
-            left_cell, left_mode = graph.base.vertices[source]
-            _right_cell, right_mode = graph.base.vertices[target]
-            add("spatial_FSWAP", "spatial_stream", left_cell, (left_mode, right_mode), (B[source], B[target], A[edge]))
-    for contact_index, (left, right) in enumerate(combinations(range(6), 2)):
-        for cell in graph.cells:
-            add(
-                "contact_phase",
-                f"contact_{contact_index}",
-                cell,
-                (left, right),
-                (bv(cell, left), bv(cell, right)),
-            )
-
-    stages = defaultdict(list)
-    for factor in factors:
-        stages[factor["stage"]].append(factor)
-    palette = total_layers = disjoint_failures = 0
-    histogram = Counter()
-    for stage in sorted(stages):
-        colors = []
-        for factor in stages[stage]:
-            for color, union in enumerate(colors):
-                if not (union & factor["support"]):
-                    colors[color] |= factor["support"]
-                    factor["color"] = color
-                    break
-            else:
-                factor["color"] = len(colors)
-                colors.append(factor["support"])
-        palette = max(palette, len(colors))
-        total_layers += len(colors)
-        histogram[len(colors)] += 1
-        by_color = defaultdict(list)
-        for factor in stages[stage]:
-            by_color[factor["color"]].append(factor["support"])
-        for masks in by_color.values():
-            union = 0
-            for mask in masks:
-                disjoint_failures += bool(union & mask)
-                union |= mask
-
-    modulus = 2 * K * length
-    maximum_diameter = max(
-        support_diameter(factor["support"], code["positions"], modulus) for factor in factors
-    )
-    cells = length**3
-    kinds = Counter(factor["kind"] for factor in factors)
-    row = {
-        "length": length,
-        "split": f"L{length}-finite-census",
-        "complete_factor_count": len(factors),
-        "expected_factor_count": 32 * cells,
-        "factor_counts": dict(kinds),
-        "ordered_stage_groups": len(stages),
-        "finite_color_palette": palette,
-        "sequential_color_layers": total_layers,
-        "stage_color_count_histogram": dict(histogram),
-        "support_disjoint_color_failures": disjoint_failures,
-        "maximum_factor_M2_weight": max(factor["weight"] for factor in factors),
-        "maximum_factor_fine_L1_diameter": maximum_diameter,
-        "constant_overhead_active_algebra_M2_per_cell": 25,
-        "factor_order_supplied": True,
-        "autonomous_controller_constructed": False,
-    }
-    row["pass"] = bool(
-        len(factors) == 32 * cells
-        and kinds
-        == {
-            "coin_Givens": 10 * cells,
-            "coin_phase": cells,
-            "reverse_FSWAP": 3 * cells,
-            "spatial_FSWAP": 3 * cells,
-            "contact_phase": 15 * cells,
-        }
-        and len(stages) == 30
-        and palette <= 7
-        and total_layers <= 58
-        and disjoint_failures == 0
-        and row["maximum_factor_M2_weight"] <= 14
-        and maximum_diameter <= 4 * K
-    )
-    return row
 
 
 def walk_symbol(momentum: np.ndarray) -> np.ndarray:
@@ -2090,7 +400,14 @@ def finite_volume_seam_controls(form_l3: np.ndarray) -> dict:
 
 
 def source_dependency_closure() -> dict:
-    source = Path(__file__).read_text(encoding="utf-8")
+    module_paths = (
+        "scripts/frontier_finite_flat_link_even_car_support_census_2026_07_23.py",
+        "scripts/frontier_finite_flat_link_even_car_support_census_graph_2026_07_23.py",
+        "scripts/frontier_finite_flat_link_even_car_support_census_factors_2026_07_23.py",
+    )
+    sources = {
+        path: (ROOT / path).read_text(encoding="utf-8") for path in module_paths
+    }
     forbidden = (
         "git" + " show",
         "merge" + "-base",
@@ -2101,29 +418,56 @@ def source_dependency_closure() -> dict:
         "cycle" + "653",
         "cycle" + "656",
     )
-    observed = {fragment: source.lower().count(fragment.lower()) for fragment in forbidden}
-    imports = tuple(
-        line.strip()
-        for line in source.splitlines()
-        if line.startswith("import ") or line.startswith("from ")
-    )
+    observed = {
+        fragment: sum(
+            source.lower().count(fragment.lower()) for source in sources.values()
+        )
+        for fragment in forbidden
+    }
+    imports = {
+        path: tuple(
+            line.strip()
+            for line in source.splitlines()
+            if line.startswith("import ") or line.startswith("from ")
+        )
+        for path, source in sources.items()
+    }
     allowed = {
         "from __future__ import annotations",
         "from collections import Counter, defaultdict",
         "from dataclasses import dataclass",
         "from hashlib import sha256",
+        "from itertools import combinations, product",
         "from itertools import combinations, permutations, product",
+        "from itertools import product",
         "import json",
         "import math",
         "from pathlib import Path",
         "import numpy as np",
         "from scipy.linalg import schur",
+        "from frontier_finite_flat_link_even_car_support_census_graph_2026_07_23 import (",
+        "from frontier_finite_flat_link_even_car_support_census_factors_2026_07_23 import (",
     }
+    unexpected = tuple(
+        sorted({line for rows in imports.values() for line in rows} - allowed)
+    )
     return {
         "imports": imports,
-        "unexpected_imports": tuple(sorted(set(imports) - allowed)),
+        "unexpected_imports": unexpected,
+        "source_modules": module_paths,
+        "source_module_bytes": {
+            path: len(source.encode("utf-8")) for path, source in sources.items()
+        },
+        "source_module_sha256": {
+            path: sha256(source.encode("utf-8")).hexdigest()
+            for path, source in sources.items()
+        },
+        "ordinary_helper_imports": AUDIT_INPUT_PATHS,
         "forbidden_reference_counts": observed,
-        "pass": not (set(imports) - allowed) and not any(observed.values()),
+        "pass": not unexpected
+        and not any(observed.values())
+        and set(AUDIT_INPUT_PATHS) == set(module_paths[1:])
+        and all(len(source) < 40_000 for source in sources.values()),
     }
